SongItemViewModelの削除
