+++ app/src/main/java/com/io/tatsuki/toney/ViewModels/SongViewModel.java
@@ -1,14 +1,26 @@
 package com.io.tatsuki.toney.ViewModels;
 
+import android.databinding.ObservableField;
+
+import com.io.tatsuki.toney.Models.Song;
+
 /**
- * Song ViewModel
+ * SongItem ViewModel
  */
 
 public class SongViewModel {
 
-    private static String TAG = SongViewModel.class.getSimpleName();
+    private static final String TAG = SongViewModel.class.getSimpleName();
+
+    public ObservableField<String> songName = new ObservableField<>();
+    public ObservableField<String> songArtist = new ObservableField<>();
 
     public SongViewModel() {
 
+    }
+
+    public void setSong(Song song) {
+        songName.set(song.getSongName());
+        songArtist.set(song.getSongArtist());
     }
 }
